// src/lib/api-zod/api.ts
export const CreateQuestionBody = zod.object({
  text: zod.string(),
  options: zod.array(zod.string()),
  correctIndex: zod.number(),
});